refactor: move _xyxy_corner_min_max to _cmc_xyxy.py, unify both CMC paths
Extract helper into trackers/core/botsort/_cmc_xyxy.py so both
apply_cmc_batch (tracker.py) and Tracklet.apply_cmc (tracklet.py) import
and call it identically — eliminates the inlined duplicate added to break
the circular import.

---
